Moon cycle decimal precision
  - Update moon cycle input to accept decimal values (step="any")
  - Change parseInt to parseFloat for cycleLength parsing
  - Set Harptos Selûne cycle to 30.4375 days (1461 days / 48 months)
  - Set Gregorian Luna cycle to 29.53059 days (Earth synodic month)

  Closes #214

+++ scripts/applications/calendar-editor.mjs
@@ -947,7 +947,7 @@ export class CalendarEditor extends HandlebarsApplicationMixin(ApplicationV2) {
       const moonColor = rawColor.toLowerCase() === '#b8b8b8' ? '' : rawColor;
       const moon = {
         name: data[`moons.${moonIdx}.name`] || '',
-        cycleLength: parseInt(data[`moons.${moonIdx}.cycleLength`]) || 28,
+        cycleLength: parseFloat(data[`moons.${moonIdx}.cycleLength`]) || 28,
         cycleDayAdjust: this.#parseOptionalInt(data[`moons.${moonIdx}.cycleDayAdjust`]) ?? existingMoon?.cycleDayAdjust ?? 0,
         color: moonColor,
         hidden: data[`moons.${moonIdx}.hidden`] === 'true' || data[`moons.${moonIdx}.hidden`] === true || existingMoon?.hidden || false,
